Add api for getting total selected row and new unsorted row
